// src/components/forms/checkbox/enum.ts
enum TagNames {
  DIV = 'div',
  LABEL = 'label',
  INPUT = 'input',
}

enum Styles {
  CONTAINER = 'checkbox-wrapper',
  INPUT_CONTAINER = 'checkbox-container',
  CHECKBOX = 'checkbox-standart',
  LABEL = 'label',
}

enum Attributes {
  TYPE = 'type',
  TYPE_VALUE_CHECKBOX = 'checkbox',
  ID = 'id',
  FOR = 'for',
}

export { TagNames, Styles, Attributes };